Move GaussianFourierProjection to seperate files to avoid import loop
