+++ experiments/eventgen/utils.py
@@ -1,0 +1,18 @@
+import torch
+import math
+from torch import nn
+
+
+class GaussianFourierProjection(nn.Module):
+    def __init__(self, embed_dim, input_dim=1, scale=30.0):
+        super().__init__()
+        # Randomly sample weights during initialization. These weights are fixed
+        # during optimization and are not trainable.
+        self.weights = nn.Parameter(
+            scale * torch.randn(input_dim, embed_dim // 2), requires_grad=False
+        )
+
+    def forward(self, t):
+        projection = 2 * math.pi * torch.matmul(t, self.weights)
+        embedding = torch.cat([torch.sin(projection), torch.cos(projection)], dim=-1)
+        return embedding
